+++ lib/features/navigation/presentation/widgets/fade_indexed_stack.dart
@@ -1,0 +1,59 @@
+import 'package:flutter/material.dart';
+
+/// Custom page transition that fades between pages
+class FadeIndexedStackTransition extends StatefulWidget {
+  final int index;
+  final List<Widget> children;
+  final Duration duration;
+
+  const FadeIndexedStackTransition({
+    super.key,
+    required this.index,
+    required this.children,
+    this.duration = const Duration(milliseconds: 300),
+  });
+
+  @override
+  State<FadeIndexedStackTransition> createState() =>
+      _FadeIndexedStackTransitionState();
+}
+
+class _FadeIndexedStackTransitionState extends State<FadeIndexedStackTransition>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _controller;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = AnimationController(
+      vsync: this,
+      duration: widget.duration,
+    );
+    _controller.forward();
+  }
+
+  @override
+  void didUpdateWidget(FadeIndexedStackTransition oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.index != oldWidget.index) {
+      _controller.forward(from: 0.0);
+    }
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return FadeTransition(
+      opacity: _controller,
+      child: IndexedStack(
+        index: widget.index,
+        children: widget.children,
+      ),
+    );
+  }
+}